requirements.txt
bcrypt==4.1.3
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.15.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Jinja2==3.1.4
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
SQLAlchemy==2.0.30
Werkzeug==3.0.3
WTForms==3.1.2
requests==2.32.2
python-dotenv==1.0.1
currency-symbols==2.0.3
email-validator==2.1.1
gunicorn==22.0.0
pytest==8.2.2